# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GPS Tracker IOT - API",
    "description": "API REST du projet GPS Tracker IOT. Permet de recevoir les positions GPS depuis un ESP32 + GPS NEO-6M, consulter l'historique, exporter les donnees et envoyer des notifications Telegram.",
    "version": "1.0.0",
    "contact": {
      "name": "Projet IOT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4900",
      "description": "Serveur local"
    }
  ],
...
  "paths": {
...
    "/api/telegram/send": {
      "post": {
        "tags": ["Telegram"],
        "summary": "Envoyer la derniere position par Telegram",
        "description": "Envoie la derniere position GPS enregistree au chat Telegram configure. Envoie un message texte avec les coordonnees suivi d'une localisation Telegram.",
        "responses": {
          "200": {
            "description": "Position envoyee avec succes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bot Telegram non configure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Telegram non configure"
                }
              }
            }
          },
          "404": {
            "description": "Aucune position disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Aucune position disponible"
                }
              }
            }
          },
          "500": {
            "description": "Erreur d'envoi Telegram",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/telegram/alert": {
      "post": {
        "tags": ["Telegram"],
        "summary": "Envoyer une alerte de geofencing",
        "description": "Envoie une alerte Telegram lorsque le tracker GPS sort d'une zone de geofencing definie. Declenchee automatiquement par le frontend.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["lat", "lng"],
                "properties": {
                  "lat": {
                    "type": "number",
                    "format": "double",
                    "description": "Latitude de la position en breach",
                    "example": -18.9150
                  },
                  "lng": {
                    "type": "number",
                    "format": "double",
                    "description": "Longitude de la position en breach",
                    "example": 47.5300
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alerte envoyee",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parametres manquants ou bot non configure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erreur d'envoi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Message d'erreur"
          }
        }
      }
    }
  }
}